# .rifle/openapi.lock.yaml
openapi: 3.1.0
info:
  title: EIFD/TIF Increment Batch API
  version: 1.0.0
  description: |
    Starts and observes the nine-step EIFD/TIF increment batch. The service uses
    preloaded shared-live assessment, tax-code, agency-reference, frozen-valuation,
    and agency-equalized-valuation data. Requests never identify local files.
    Every launch pins the business clock for the complete chain. Runs share the
    live indexed stores and are not isolated snapshots.
servers:
- url: http://localhost:8080
  description: Local development
...
security: []
paths:
  /api/assessed-value-preparation-runs:
    post:
      tags:
      - Assessed Value Preparation Runs
      summary: Start an assessed value preparation run
      operationId: startAssessedValuePreparationRun
      description: |
        Creates an asynchronous run over the assessment data already preloaded
        on the server. The supplied business date and time are pinned for every
        stage, and `processYear` supplies the evidenced two-character ASREA178
        parameter. The accepted business rules below all apply at this endpoint.
        Their final batch outcome, counts, outputs, and messages are observed by
        reading the returned resource through
        `GET /api/assessed-value-preparation-runs/{id}`.

        Failure and partial-output semantics are explicit. A descending key in
        overall-class preparation produces a `FAILED` run, return code 16, an
        `asrea018-001` error message, and `partialOutput: true` when records were
        published. A descending key in valuation bucketing also produces a
        `FAILED` run and `partialOutput: true`, even though the legacy stage can
        report return code zero; its message names `asrea151-001`. Outputs made
        before either failure remain listed with `partial: true` and are never
        represented as a clean successful result. Source defects such as invalid
        decimal selectors, major-9 switch contamination, and nonpositive Type-5
        occupancy are likewise surfaced in rule-addressed messages.
      x-business-rules:
      - id: asrea003-001
        behavior: Value one supplied supported land or improvement detail without independently reading or publishing a batch dataset.
      - id: asrea003-002
        behavior: Zero EX/RR land; otherwise apply supported decimal scaling, unit price, positive adjustment factors, and the one-dollar minimum.
      - id: asrea003-003
        behavior: Preserve the observed one-dollar result for decimal selectors 6 through 9 and emit a defect message naming this rule.
      - id: asrea003-004
        behavior: Value Type-2 area improvements from scaled area, unit price, condition, optional positive assessed percentage, and the one-dollar minimum.
      - id: asrea003-005
        behavior: Apply class-288 reproduction-cost relief using the accepted year bands and thresholds.
      - id: asrea003-006
        behavior: Start Type-3 valuation from reproduction cost; start Type-4/5 only for positive condition, while preserving the documented zero and one-dollar boundaries.
      - id: asrea018-001
        behavior: Require nondecreasing township, volume, parcel, and tax-type order and expose failed partial publication explicitly.
      - id: asrea018-002
        behavior: Use the highest land class for land-only parcels and report parcels that contain no land detail.
      - id: asrea018-003
        behavior: Resolve named improvement classes by the fixed first-match major-class precedence.
      - id: asrea018-004
        behavior: Resolve other recognized improvement classes by major class, tier priority, valuation, and minor class, neutralizing minor class 88.
      - id: asrea018-005
        behavior: Preserve the prior overall class and report the record when improvement classification cannot be resolved.
      - id: asrea018-006
        behavior: Exclude attached questionnaire occurrences from overall-class selection.
      - id: asrea018-007
        behavior: Identify observed major-9 tier-2 contamination of the major-8 switch as a reported source defect, not an endorsed classification policy.
      - id: asrea151-001
        behavior: Require nondecreasing township, volume, and parcel order and expose the legacy zero-return-code partial output as a failed run.
      - id: asrea151-002
        behavior: Map overall Assessor class through the 242-entry table; report nonzero misses and use the accepted major-digit fallback.
      - id: asrea151-003
        behavior: Establish parcel bucket eligibility using class-239/class-200 land and the strict residential and class-299 value thresholds.
      - id: asrea151-004
        behavior: Accumulate class-239 land and class-224 improvements into farm value when class-239 land is present.
      - id: asrea151-005
        behavior: Accumulate eligible major-class-2 non-farm detail into homeowner value and all remaining detail into non-homeowner value.
      - id: asrea151-006
        behavior: Publish three distinct report buckets while persisting farm separately, homeowner plus non-homeowner combined, and the documented cleared slots.
      - id: asrea151-007
        behavior: Exclude questionnaire occurrences from both valuation-bucket passes.
      - id: asrea178-001
        behavior: Require a numeric two-character process year and use the accepted pivot only for report-century display.
      - id: asrea178-002
        behavior: Clear Type-2 CDU year and exclude attached questionnaires from conversion scanning.
      - id: asrea178-003
        behavior: Report nonpositive Type-5 occupancy as a source defect while recording the observed conversion without cost gross-up.
      - id: asrea178-004
        behavior: Normalize positive-occupancy Type-5 reproduction cost and map the detail to Type 3, code 2, and year zero.
      - id: asrea178-005
        behavior: Revalue every non-questionnaire detail under ASREA003 policy when any Type-5 detail is found.
      - id: asrea178-006
        behavior: Separate recomputed proposed land and improvement totals by detail type while skipping questionnaires.
      - id: asrea178-007
        behavior: Archive the prior proposed total in slot 12 and replace proposed land, improvement, and total values.
      - id: asrea178-008
        behavior: Do not apply the unreachable legacy township lookup as an active validation guard.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessedValuePreparationRunRequest'
            examples:
              standardRun:
                value:
                  businessDate: 2025-09-15
                  businessTime: 12:00:00
                  idempotencyKey: avp-2025-09-15-noon
                  processYear: '26'
      responses:
        '201':
          description: Run created and queued or already running asynchronously.
          headers:
            Location:
              description: URI of the created run resource.
              schema:
                type: string
                example: /api/assessed-value-preparation-runs/841732905
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessedValuePreparationRunResponse'
              examples:
                queued:
                  value:
                    id: 841732905
                    status: QUEUED
                    businessDate: 2025-09-15
                    businessTime: 12:00:00
                    processYear: '26'
                    outputs: []
                    messages: []
        '400':
          description: The run request is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessedValuePreparationErrorResponse'
              examples:
                invalidProcessYear:
                  value:
                    error: INVALID_PROCESS_YEAR
                    message: processYear must contain exactly two decimal digits.
                    ruleId: asrea178-001
        '409':
          description: The idempotency key is already associated with a different assessed-value preparation request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessedValuePreparationErrorResponse'
              examples:
                idempotencyConflict:
                  value:
                    error: IDEMPOTENCY_CONFLICT
                    message: The idempotencyKey is already associated with different request parameters.
  /api/assessed-value-preparation-runs/{id}:
    get:
      tags:
      - Assessed Value Preparation Runs
      summary: Read an assessed value preparation run
      operationId: getAssessedValuePreparationRun
      description: |
        Returns operational status and the latest observable batch result. This
        endpoint does not define business decisions; all business rules attach to
        `POST /api/assessed-value-preparation-runs`. When processing has ended,
        this response is where callers observe the final return code, record
        counts, output artifacts, stage results, partial-output flag, and
        rule-addressed messages.
      parameters:
      - name: id
        in: path
        required: true
        description: Positive run identifier returned when the run was created.
        schema:
          type: integer
          format: int64
          minimum: 1
        example: 841732905
      responses:
        '200':
          description: Current status and result of the run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessedValuePreparationRunResponse'
              examples:
                completed:
                  value:
                    id: 841732905
                    status: COMPLETED
                    businessDate: 2025-09-15
                    businessTime: 12:00:00
                    processYear: '26'
                    returnCode: 0
                    recordsRead: 30
                    recordsWritten: 30
                    recordsUpdated: 10
                    recordsRejected: 0
                    partialOutput: false
                    outputs:
                    - artifactId: avpout_01K5A8C2V7M4R9N6T3Q0W1X8YZ
                      kind: OVERALL_CLASS_REPORT
                      mediaType: text/plain
                      recordCount: 14
                      partial: false
                    - artifactId: avpout_01K5A8D4B2H7P9S6F3J0L1M5NC
                      kind: VALUATION_BREAKDOWN_REPORT
                      mediaType: text/plain
                      recordCount: 17
                      partial: false
                    - artifactId: avpout_01K5A8F6D3N9R2T7V4X0Z1B8GH
                      kind: ASSESSMENT_MASTER
                      mediaType: application/octet-stream
                      recordCount: 10
                      partial: false
                    messages: []
                    stages:
                    - stage: OVERALL_CLASS
                      status: COMPLETED
                      returnCode: 0
                      recordsRead: 10
                      recordsWritten: 10
                      recordsUpdated: 10
                      recordsRejected: 0
                      outputPublished: true
                      partialOutput: false
                    - stage: VALUATION_BUCKETING
                      status: COMPLETED
                      returnCode: 0
                      recordsRead: 10
                      recordsWritten: 10
                      recordsUpdated: 10
                      recordsRejected: 0
                      outputPublished: true
                      partialOutput: false
                    - stage: TYPE5_CONVERSION
                      status: COMPLETED
                      returnCode: 0
                      recordsRead: 10
                      recordsWritten: 10
                      recordsUpdated: 0
                      recordsRejected: 0
                      outputPublished: true
                      partialOutput: false
        '404':
          description: No run exists for the supplied identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessedValuePreparationErrorResponse'
              examples:
                runNotFound:
                  value:
                    error: RUN_NOT_FOUND
                    message: No assessed-value preparation run exists for the supplied id.
...
components:
  schemas:
    AssessedValuePreparationRunRequest:
      type: object
      additionalProperties: false
      required:
      - businessDate
      - businessTime
      - idempotencyKey
      - processYear
      properties:
        businessDate:
          type: string
          format: date
          description: Business date pinned for the entire run.
        businessTime:
          type: string
          pattern: ^(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$
          description: Business time pinned for the entire run in HH:mm:ss form.
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 128
          description: Caller key that prevents duplicate creation of the same batch request.
        processYear:
          type: string
          pattern: ^[0-9]{2}$
          description: Numeric two-character Type-5 conversion process year (asrea178-001).
    AssessedValuePreparationRunResponse:
      type: object
      additionalProperties: false
      required:
      - id
      - status
      - businessDate
      - businessTime
      - processYear
      - outputs
      - messages
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: Positive server-generated run identifier.
        status:
          type: string
          enum:
          - QUEUED
          - RUNNING
          - COMPLETED
          - FAILED
        businessDate:
          type: string
          format: date
        businessTime:
          type: string
          pattern: ^(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$
        processYear:
          type: string
          pattern: ^[0-9]{2}$
        returnCode:
          type: integer
          description: Final or latest legacy return code when one is available. Status remains FAILED for a detected rule failure even when a legacy stage left this value at zero.
        recordsRead:
          type: integer
          minimum: 0
        recordsWritten:
          type: integer
          minimum: 0
        recordsUpdated:
          type: integer
          minimum: 0
        recordsRejected:
          type: integer
          minimum: 0
        partialOutput:
          type: boolean
          description: True when any listed output was published from an incomplete or failed transformation, including asrea018-001 and asrea151-001.
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/AssessedValuePreparationOutput'
        messages:
          type: array
          description: Business warnings and failures, including preserved legacy defects, addressed by exact accepted rule ID.
          items:
            $ref: '#/components/schemas/AssessedValuePreparationMessage'
        stages:
          type: array
          items:
            $ref: '#/components/schemas/AssessedValuePreparationStageResult'
    AssessedValuePreparationOutput:
      type: object
      additionalProperties: false
      required:
      - artifactId
      - kind
      - mediaType
      - recordCount
      - partial
      properties:
        artifactId:
          type: string
          minLength: 20
          maxLength: 128
          pattern: ^[A-Za-z0-9_-]+$
          description: Server-managed artifact identifier; never a local filesystem path.
        kind:
          type: string
          enum:
          - OVERALL_CLASS_REPORT
          - VALUATION_BREAKDOWN_REPORT
          - ASSESSMENT_MASTER
          - TYPE5_ERROR_REPORT
        mediaType:
          type: string
        recordCount:
          type: integer
          minimum: 0
        partial:
          type: boolean
          description: True when this artifact was published despite an incomplete or failed run.
    AssessedValuePreparationMessage:
      type: object
      additionalProperties: false
      required:
      - severity
      - code
      - message
      properties:
        severity:
          type: string
          enum:
          - INFO
          - WARNING
          - ERROR
        code:
          type: string
          minLength: 1
        ruleId:
          type: string
          pattern: ^asrea(?:003|018|151|178)-[0-9]{3}$
          description: Exact accepted rule ID required for a business-rule message; omitted for an operational worker or input failure.
        message:
          type: string
          minLength: 1
        recordKey:
          type: string
          description: Server-rendered parcel key when the message concerns one record.
    AssessedValuePreparationStageResult:
      type: object
      additionalProperties: false
      required:
      - stage
      - status
      - recordsRead
      - recordsWritten
      - recordsUpdated
      - recordsRejected
      - outputPublished
      - partialOutput
      properties:
        stage:
          type: string
          enum:
          - OVERALL_CLASS
          - VALUATION_BUCKETING
          - TYPE5_CONVERSION
        status:
          type: string
          enum:
          - NOT_STARTED
          - RUNNING
          - COMPLETED
          - FAILED
        returnCode:
          type: integer
          description: Stage return code when available; may be zero on failed ASREA151 ordering semantics.
        recordsRead:
          type: integer
          minimum: 0
        recordsWritten:
          type: integer
          minimum: 0
        recordsUpdated:
          type: integer
          minimum: 0
        recordsRejected:
          type: integer
          minimum: 0
        outputPublished:
          type: boolean
        partialOutput:
          type: boolean
          description: Explicit stage-level partial publication indicator.
    AssessedValuePreparationErrorResponse:
      type: object
      additionalProperties: false
      required:
      - error
      - message
      properties:
        error:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        ruleId:
          type: string
          description: Exact accepted rule ID when a request error is governed by a business rule; omitted for operational errors.
...